Scotts Changes
I've added a css sheet, and used it to center the elements that were
previously inline centered. I have added a basic color scheme to the
page, and added a comment containing the code to make the page a
dragscroll element. Finally I have moved a few peices of code around to
make the css work, but have only removed the aforementioned <center>
blocks. (Side note, I changed the ID of the main divbox so it is more
clearly labeled. )